agregar endpoint

+++ server/House.py
@@ -1,8 +1,8 @@
 from pydantic import BaseModel
 
 class House(BaseModel):
-    bedrooms: int | None = None
+    bedrooms: float | None = None
     bathrooms: float | None = None
-    area: int | None = None
+    area: float | None = None
     Latitud: float | None = None
     Longitud: float | None = None
